Move template function implementation to fix linking errors
